Give the size options one type and the two bare integers their enums
Items 6 and 7 of #125.

**Twenty size members are ulong rather than nuint.** The same native size_t was
ulong on some members and nuint on others: WriteBufferSize against
ArenaBlockSize, Cache.CreateLru against WriteBufferManager.Create,
HyperClockCacheOptions taking ulong and exposing nuint. nuint is also
platform-dependent in a public API, which is not a property a size option
should have. The conversion sits at the P/Invoke boundary as
checked((nuint)value), so on a 32-bit process a value above uint.MaxValue
throws rather than truncating to something smaller than asked for.

The generated layer is untouched and stays nuint, because that is what
size_t is. This was a wrapper problem, not a generator one.

**BlockBasedTableOptions.Checksum is the ChecksumType enum.** It was an int on
the grounds that the C API does not publish the values. It does not, but the
header does: ChecksumType in include/rocksdb/table.h at the pinned version, five
members, and the numbering cannot drift because the value is written into every
SST file and read back by whichever version opens it later.

Checked against the running library as well as the header. RocksDb accepts
exactly 0 to 4 and rejects anything outside, naming ChecksumType when it does; a
fresh options object reports 4; and RocksDb's own option parser accepts exactly
the five names the members are called after, and rejects an invented one. Three
tests pin those.

**DbOptions.VerifyOutputFlags is a [Flags] enum.** Described as flags RocksDb
does not expose, and it does: VerifyOutputFlags in
include/rocksdb/advanced_options.h, seven members. The interesting part is the
shape, which nothing said: the bits come in two groups, three saying what to
verify and two saying which compactions to verify it on, and a value needs one
from each group to do anything. That is why setting a single bit had no
observable effect. The C API takes and returns int against RocksDb's uint32_t,
which is why All arrives as -1, and why the old int property accepted -1 as a
value: that was asking for everything.

Not done, so #125 stays open: the ticker and histogram ids that GetTickerCount
and GetHistogramData take as bare numbers. Measured in statistics.h at the
pinned version, those are 263 and 80 members with almost entirely implicit,
positional values, and they gain and lose entries between releases. Hand-writing
343 members that renumber when RocksDb reorders one is the wrong shape; the
generator already fetches c.h at the pinned tag and should fetch statistics.h
the same way. That is its own change.

The changelog gains eleven rows, covering these three and the five items merged
in #134, which never made it into the migration table the README points at. The
row for WriteBatchWithIndex.DeleteRange described it as throwing and obsolete
and is now correct about it being gone.

1054 tests pass on net8.0, net9.0 and net10.0. Build clean under -warnaserror
from a clean build.

+++ RocksDb.Net.Tests/TransactionOptionsTests.cs
@@ -41,7 +41,7 @@ public class TransactionOptionsTests
         Assert.Equal(1234, opts.TransactionLockTimeout);
         Assert.Equal(5678, opts.DefaultLockTimeout);
         Assert.Equal(999, opts.MaxNumLocks);
-        Assert.Equal((nuint)32, opts.NumStripes);
+        Assert.Equal((ulong)32, opts.NumStripes);
         Assert.Equal(7u, opts.MaxNumDeadlocks);
         Assert.Equal(4096, opts.DefaultWriteBatchFlushThreshold);
         Assert.Equal(8192u, opts.CommitBypassMemtableThreshold);
@@ -84,7 +84,7 @@ public class TransactionOptionsTests
         Assert.Equal(64, opts.DeadlockDetectDepth);
         Assert.Equal(250, opts.LockTimeout);
         Assert.Equal(30_000, opts.Expiration);
-        Assert.Equal((nuint)(1 << 20), opts.MaxWriteBatchSize);
+        Assert.Equal((ulong)(1 << 20), opts.MaxWriteBatchSize);
         Assert.Equal(1500, opts.DeadlockTimeoutMicros);
         Assert.Equal(2048, opts.WriteBatchFlushThreshold);
         Assert.Equal(100u, opts.LargeTransactionCommitOptimizeThreshold);
